+++ layers/GatoGraphQLForWP/plugins/gatographql/src/PluginInitializationConfiguration.php
@@ -10,6 +10,7 @@ use GatoGraphQL\GatoGraphQL\Constants\ModuleSettingOptions;
 use GatoGraphQL\GatoGraphQL\Facades\Instances\PluginOptionsFormHandlerFacade;
 use GatoGraphQL\GatoGraphQL\Facades\Registries\SystemModuleRegistryFacade;
 use GatoGraphQL\GatoGraphQL\Facades\UserSettingsManagerFacade;
+use GatoGraphQL\GatoGraphQL\ModuleResolvers\CachingFunctionalityModuleResolver;
 use GatoGraphQL\GatoGraphQL\ModuleResolvers\ClientFunctionalityModuleResolver;
 use GatoGraphQL\GatoGraphQL\ModuleResolvers\EndpointFunctionalityModuleResolver;
 use GatoGraphQL\GatoGraphQL\ModuleResolvers\MetaSchemaTypeModuleResolver;
@@ -60,6 +61,8 @@ use PoPCMSSchema\UserRoles\Environment as UserRolesEnvironment;
 use PoPCMSSchema\UserRoles\Module as UserRolesModule;
 use PoPCMSSchema\Users\Environment as UsersEnvironment;
 use PoPCMSSchema\Users\Module as UsersModule;
+use PoPAPI\API\Environment as APIEnvironment;
+use PoPAPI\API\Module as APIModule;
 use PoPSchema\Logger\Environment as LoggerEnvironment;
 use PoPSchema\Logger\Module as LoggerModule;
 use PoPSchema\SchemaCommons\Constants\Behaviors;
@@ -108,6 +111,13 @@ class PluginInitializationConfiguration extends AbstractMainPluginInitialization
                 'envVariable' => ComponentModelEnvironment::CLIENT_IP_ADDRESS_SERVER_PROPERTY_NAME,
                 'module' => PluginGeneralSettingsFunctionalityModuleResolver::SERVER_IP_CONFIGURATION,
                 'option' => PluginGeneralSettingsFunctionalityModuleResolver::OPTION_CLIENT_IP_ADDRESS_SERVER_PROPERTY_NAME,
+            ],
+            // Cache parsed GraphQL queries?
+            [
+                'class' => APIModule::class,
+                'envVariable' => APIEnvironment::USE_PARSED_AST_CACHE,
+                'module' => CachingFunctionalityModuleResolver::CACHING,
+                'option' => CachingFunctionalityModuleResolver::OPTION_USE_PARSED_AST_CACHE,
             ],
             // Editing Access Scheme
             [